requirement.txt
asgiref
beautifulsoup4
certifi
chardet
defusedxml
Django
django-allauth
django-appconf
django-ckeditor
django-classy-tags
django-disqus
django-js-asset
django-mptt
django-reversion
django-taggit
djangorestframework
idna
lxml
mock
mysqlclient
oauthlib
python3-openid
pytz
requests
requests-oauthlib
six
soupsieve
sqlparse
urllib3
django-taggit
mysqlclient
